refactor(datacollection(): reorganize files
