+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/List/components/Row.tsx
@@ -13,10 +13,10 @@ import {
 } from "../../../../item-actions"
 import { ItemActionsDropdown } from "../../../../ItemActions/ItemActionsDropdown"
 import { NavigationFiltersDefinition } from "../../../../navigationFilters/types"
-import { renderProperty } from "../../../../property-render"
 import { SortingsDefinition } from "../../../../sortings"
 import { SummariesDefinition } from "../../../../summary"
 import { DataSource, GroupingDefinition, RecordType } from "../../../../types"
+import { renderProperty } from "../../../property/property-render"
 import { actionsToDropdownItems } from "../../utils"
 import { ItemDefinition, ListPropertyDefinition } from "../types"
 import { ItemTeaser } from "./ItemTeaser"
